Apply metrics monitoring middlewares
