+++ server/src/monitor/index.ts
@@ -1,4 +1,5 @@
 export * from './logger';
+export * from './metrics';
 export * from './prometheus';
-export * from './zipkin';
 export * from './telemetry';
+export * from './zipkin';
